pyproject.toml
[tool.poetry]
name = "pet-shop-api"
version = "0.1.0"
description = "API для магазина домашних животных"
authors = ["Larisa Shirokikh <lorashirokikh@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.0"
uvicorn = "^0.23.2"
sqlalchemy = "^2.0.22"
alembic = "^1.12.0"
pydantic = "^2.4.2"
pydantic-settings = "^2.0.3"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.6"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
asyncpg = "^0.28.0"
psycopg2-binary = "^2.9.9"
email-validator = "^2.0.0"
tenacity = "^8.2.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-asyncio = "^0.21.1"
httpx = "^0.25.0"
black = "^23.10.0"
isort = "^5.12.0"
mypy = "^1.6.1"
flake8 = "^6.1.0"
pre-commit = "^3.5.0"
alembic = "^1.15.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
line_length = 88

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false